requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-restful
marshmallow
marshmallow-sqlalchemy
pytest
pymysql
Flask-JWT
bcrypt
